+++ examples/next-typescript-starter/app/api/custom-claims/route.ts
@@ -16,8 +16,6 @@ export async function POST(request: NextRequest) {
     throw new Error('Cannot update custom claims of unauthenticated user');
   }
 
-  const appCheckToken = request.headers.get('X-Firebase-AppCheck') ?? undefined;
-
   await setCustomUserClaims(tokens.decodedToken.uid, {
     someCustomClaim: {
       updatedAt: Date.now()
@@ -39,8 +37,5 @@ export async function POST(request: NextRequest) {
     }
   );
 
-  return refreshNextResponseCookies(request, response, {
-    ...authConfig,
-    appCheckToken
-  });
+  return refreshNextResponseCookies(request, response, authConfig);
 }