+++ src/index.ts
@@ -4,7 +4,11 @@ export {
   redirectToLogin
 } from './next/middleware';
 
-export {getTokens, getTokensFromObject} from './next/tokens';
+export {
+  getTokens,
+  getTokensFromObject,
+  getApiRequestTokens
+} from './next/tokens';
 
 export {getFirebaseAuth} from './auth';
 export type {Tokens} from './auth';